Improve step interval range handling; use "starting" prefix
This introduces a new resource called CommStartingX0 which is used for
step interval range expressions (i.e. 2/3) so that that, for example,
the expression "5/30 * * * * ?" reads as:

"Every 30 seconds, starting at 5 seconds past the minute"

whereas before it read as:

"Every 30 seconds, seconds 5 through 59 past the minute"

Previously, we were converting all step interval ranges to "between"
expressions to be able to re-use the existing translations but the
descriptions were awkward and misleading.

We will leave the "between" conversion in place for Month, DOW, and Year
 parts as between handling results in acceptable descriptions and avoids
more significant reworking of the translations.